Velocities somewhat working
Velocity interpolation seems to work ok for a grid that is full, i.e. not a dugeon. I've added grid cells in obstructed areas so that there aren't null pointers in the interpolation method however this doesn't work in all cases. Assuming that the physics collisions requires the agents to stay in the unobstructed cells it should work.