Working Slime Crowds
They get stuck in halways, but it's good for testing in large rooms (for seeing them interact with placed objects etc). Just click perform to create a dungeon, then initialize crowd simulation, then add a goal to the dugeon, then click add agent button to add agents to the simulation.