Refactoring Interface
Created a gameObject for the simulation. Now the simAccess object can be set in the proceduralPipeline object.

Pressing perform will generate a dungeon and a crowd sim attatched to it. Continusously pressing perform will continue todo so.

You can click initialize grid, without clicking perform, which will just create a plane grid for the crowd sim (for testing).